Crate twoway

Modules

bmh

Boyer-Moore-Horspool

Functions

find_bytes

find_bytes finds the first ocurrence of pattern in the text.

find_str

find_str finds the first ocurrence of pattern in the text.

rfind_bytes

rfind_bytes finds the last ocurrence of pattern in the text, and returns the index of the start of the match.

rfind_str

rfind_str finds the last ocurrence of pattern in the text and returns the index of the start of the match.